Set the folder organization, load of graphics assets
Managed the folders in a way everyone can use.
Loaded assets for UI, Inventory, Health and Menu.
Loaded all the assets for the graphic of the Overgrown Forest, plus some others for the ruins, the wastelands and the l;aboratory. First scene of the overgrown forest still very primordial.
Note: the Sprite for Wollo is just for demo purposes.